Replaced mxinterp with gslCubSplineInterp_UB.
This is NOT the final version, it is just an intermediate step!

The number of threads is mostly hard-coded and set to 4.
The number of threads is read from the parameters file for xinterp_dm only:
I added a member "interpolationNthreads" to classes "dm" and "MirrorSurface".
For number of threads = 0, the interpolation will be executed as a function call without creation of a thread. For "default" number of threads, -1 should be used.

Two minor changes to Ctrl.py:
(1) change a name of a variable in a short for-loop:
    for c in compList
    ==>
    for module in compList
(2) At the end of simulation, print out the total time spent at each component.