(1) renamed the following two functions throughout aosim: -  gslCubSplineInterp    ==> gslPeriodicCubSplineInterp -  gslCubSplineInterp_UB ==> gslCubSplineInterp
(2) Replaced the hard-coded number of threads in calls of gslCubSplineInterp
with a parameter that is read from the parameters file.
Added a new member "interpolationNthreads" to classes "infAtmos", "atmos", "vdmUser" and "dm" (from zdm.py).
I was not able to test the code after the changes fully, because in my current simulation not all of these modules are used.